backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
anthropic
pdfplumber
PyMuPDF
python-docx
pytesseract
Pillow
weasyprint
markdown
python-dotenv
pydantic-settings
python-multipart
pydantic[email]
